fix: circular dependencies in type imports
