+++ src/models/user_pp.models.tsx
@@ -1,6 +1,6 @@
 import type { IUser } from './user.models'
 import type { DBDoc } from './common.models'
-import type { ProfileTypeLabel } from '../modules/profile'
+import type { ProfileTypeLabel } from '../modules/profile/types'
 
 export type PlasticTypeLabel =
   | 'pet'
