+++ src/modules/profile/index.ts
@@ -2,15 +2,6 @@ import { getConfigurationOption } from '../../config/config'
 import { SupportedProfileTypesFactory } from './SupportedProfileTypesFactory'
 import type { PlatformTheme } from 'src/themes/types'
 
-export const ProfileType = {
-  MEMBER: 'member',
-  SPACE: 'space',
-  WORKSPACE: 'workspace',
-  MACHINE_BUILDER: 'machine-builder',
-  COMMUNITY_BUILDER: 'community-builder',
-  COLLECTION_POINT: 'collection-point',
-} as const
-
 export function getSupportedProfileTypes(currentTheme?: PlatformTheme) {
   const supportedProfileTypes = SupportedProfileTypesFactory(
     getConfigurationOption('REACT_APP_PLATFORM_PROFILES', ''),
@@ -19,7 +10,3 @@ export function getSupportedProfileTypes(currentTheme?: PlatformTheme) {
 
   return supportedProfileTypes
 }
-
-export type ProfileTypeLabel = typeof ProfileType[keyof typeof ProfileType]
-
-export type { IProfileType } from './types'
