+++ src/modules/profile/types.ts
@@ -1,4 +1,13 @@
-import type { ProfileTypeLabel } from './index'
+export const ProfileType = {
+  MEMBER: 'member',
+  SPACE: 'space',
+  WORKSPACE: 'workspace',
+  MACHINE_BUILDER: 'machine-builder',
+  COMMUNITY_BUILDER: 'community-builder',
+  COLLECTION_POINT: 'collection-point',
+} as const
+
+export type ProfileTypeLabel = typeof ProfileType[keyof typeof ProfileType]
 
 export interface IProfileType {
   label: ProfileTypeLabel
